+++ packages/vite/src/executors/dev-server/dev-server.impl.ts
@@ -77,6 +77,10 @@ export async function* viteDevServerExecutor(
 
 async function runViteDevServer(server: ViteDevServer): Promise<void> {
   await server.listen();
+  process.send({
+    type: 'nx.server.ready',
+    baseUrl: server.resolvedUrls.local[0],
+  });
   server.printUrls();
 
   const processOnExit = async () => {
